more assets

+++ Game/EntityManager.h
@@ -81,7 +81,7 @@ void* GetAsset(char* _file);
 
 sfSprite* LoadBackground(sfTexture* _texture, float _scale);
 sfSprite* CreateSprite(sfTexture* _texture, sfVector2f _position, float _scale, float _drawPlan);
-sfText* CreateText(sfFont* _font, sfVector2f _position, float _scale, float _drawPlan);
+sfText* CreateText(sfFont* _font, sfVector2f _position, unsigned _scale, float _drawPlan);
 sfSound* CreateSound(sfSoundBuffer* _buffer, float _volume, sfBool _play);
 sfMusic* CreateMusic(char* _fileMusic, float _volume, sfBool _play);
